+++ MatrixClasses/RealMatrix.h
@@ -9,8 +9,8 @@
  *
  * @version     kspaceFirstOrder3D 2.15
  *
- * @date        11 July 2011, 10:30      (created) \n
- *              20 June 2014, 15:17 (revised)
+ * @date        11 July      2011, 10:30 (created) \n
+ *              01 September 2014, 13:12 (revised)
  *
  * @section License
  * This file is part of the C++ extension of the k-Wave Toolbox (http://www.k-wave.org).\n
@@ -35,7 +35,6 @@
 
 
 #include <MatrixClasses/BaseFloatMatrix.h>
-#include <MatrixClasses/LongMatrix.h>
 
 #include <Utils/DimensionSizes.h>
 
@@ -53,7 +52,10 @@ class TRealMatrix : public TBaseFloatMatrix
     TRealMatrix(struct TDimensionSizes DimensionSizes);
 
      /// Destructor
-    virtual ~TRealMatrix() { FreeMemory(); };
+    virtual ~TRealMatrix()
+    {
+      FreeMemory();
+    };
 
     /// Read data from the HDF5 file - only from the root group
     virtual void ReadDataFromHDF5File(THDF5_File & HDF5_File,
@@ -62,7 +64,7 @@ class TRealMatrix : public TBaseFloatMatrix
     /// Write data into the HDF5 file
     virtual void WriteDataToHDF5File(THDF5_File & HDF5_File,
                                      const char * MatrixName,
-                                     const int CompressionLevel);
+                                     const size_t CompressionLevel);
 
     /**
      * @brief operator []
